waste_management_using_cnn/requirements.txt
torch==2.1.0
opencv-python==4.6.0.66
Pillow==8.2.0
torchvision==0.15.0
ultralytics==8.0.0
